fix: pin major version of open-next to 1
Currently, `cdk-nextjs` is not compatible with the latest v2+ version of `open-next`. Work is being done to update the project to be compatible, but for now we'll make sure new users coming to `cdk-nextjs` for the first time are to deploy a working Next.js site.

see: https://github.com/jetbridge/cdk-nextjs/issues/126