revert default origin to original behavior
I now see why this was in place originally. It still doesn't work, but I'd rather leave it as it is for now.

SST creates additional behaviors in CloudFormation at deploy time, basically iterating through all of the files in the root of `/public` and creating a rule that routes to the static assets bucket for these matches:
https://github.com/serverless-stack/sst/blob/b56c2ea021290211c72841c605cec58579ef3591/packages/sst/src/constructs/SsrSite.ts#L1113
```
  protected addStaticFileBehaviors() {
    const { cdk } = this.props;

    // Create a template for statics behaviours
    const publicDir = path.join(
      this.props.path,
      this.buildConfig.clientBuildOutputDir
    );
    for (const item of fs.readdirSync(publicDir)) {
      const isDir = fs.statSync(path.join(publicDir, item)).isDirectory();
      this.distribution.addBehavior(isDir ? `${item}/*` : item, this.s3Origin, {
        viewerProtocolPolicy: ViewerProtocolPolicy.REDIRECT_TO_HTTPS,
        allowedMethods: AllowedMethods.ALLOW_GET_HEAD_OPTIONS,
        cachedMethods: CachedMethods.CACHE_GET_HEAD_OPTIONS,
        compress: true,
        cachePolicy: CachePolicy.CACHING_OPTIMIZED,
        responseHeadersPolicy: cdk?.responseHeadersPolicy,
      });
    }
  }
```

This is a bit more complicated with the pitfall of hitting the maximum number of behaviors (25 apparently).

> First do no evil.