bugfix: fix XAmzContentSHA256Mismatch (#249)
This appears to be a bug in the AWS s3 SDK. The root of the issue seems to be calling upload on lib-storage rather than an issue with the implementation. Switching from promise.all to a serial implentation fixes the issue, while also increasing the queue size off the default 3.

For future improvements, I would also thing about exposing more of the uploadOptions inside the next deployment construct. Perhaps we should also give consumers the options of uploading serial vs parallel.

Next steps I will also raise this with AWS team.